Crate rustc_monomorphize

Modules

collector 🔒

Mono Item Collection

errors 🔒

mono_checks 🔒

This implements a single query, check_mono_fn, that gets fired for each monomorphization of all functions. This lets us implement monomorphization-time checks in a way that is friendly to incremental compilation.

partitioning 🔒

Partitioning Codegen Units for Incremental Compilation

util 🔒

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Functions

custom_coerce_unsize_info 🔒

provide