rustc_monomorphize

Module mono_checks

This implements a single query, check_mono_fn, that gets fired for each monomorphization of all functions. This lets us implement monomorphization-time checks in a way that is friendly to incremental compilation.

Modules

abi_check 🔒

This module ensures that if a function’s ABI requires a particular target feature, that target feature is enabled both on the callee and all callers.

move_check 🔒

Functions

check_mono_item 🔒

provide 🔒