rustc_monomorphize::mono_checks

Module abi_check

This module ensures that if a function’s ABI requires a particular target feature, that target feature is enabled both on the callee and all callers.

Functions

check_call_site_abi 🔒

Checks that a call expression does not try to pass a vector-passed argument which requires a target feature that the caller does not have, as doing so causes UB because of ABI mismatch.

check_callees_abi 🔒

check_feature_dependent_abi 🔒

check_instance_abi 🔒

Checks that the ABI of a given instance of a function does not contain vector-passed arguments or return values for which the corresponding target feature is not enabled.

do_check_simd_vector_abi 🔒

Checks whether a certain function ABI is compatible with the target features currently enabled for a certain function. is_call indicates whether this is a call-site check or a definition-site check; this is only relevant for the wording in the emitted error.

uses_vector_registers 🔒