rustc_monomorphize::mono_checks::abi_check
Function do_check_abi

fn do_check_abi<'tcx>(
    tcx: TyCtxt<'tcx>,
    abi: &FnAbi<'tcx, Ty<'tcx>>,
    target_feature_def: DefId,
    emit_err: impl FnMut(Option<&'static str>),
)

Checks whether a certain function ABI is compatible with the target features currently enabled for a certain function. If not, emit_err is called, with Some(feature) if a certain feature should be enabled and with None if no feature is known that would make the ABI compatible.