rustc_monomorphize::partitioning
Function dump_mono_items_stats

fn dump_mono_items_stats<'tcx>(
    tcx: TyCtxt<'tcx>,
    codegen_units: &[CodegenUnit<'tcx>],
    output_directory: &Option<PathBuf>,
    crate_name: Symbol,
) -> Result<(), Box<dyn Error>>

Outputs stats about instantiation counts and estimated size, per MonoItem’s def, to a file in the given output directory.