rustc_monomorphize::polymorphize
Function mark_used_by_default_parameters

fn mark_used_by_default_parameters<'tcx>(
    tcx: TyCtxt<'tcx>,
    def_id: DefId,
    generics: &'tcx Generics,
    unused_parameters: &mut UnusedGenericParams,
)

Some parameters are considered used-by-default, such as non-generic parameters and the dummy generic parameters from closures, this function marks them as used. leaf_is_closure should be true if the item that unused_generic_params was invoked on is a closure.