rustc_monomorphize::util
Function dump_closure_profile

pub(crate) fn dump_closure_profile<'tcx>(
    tcx: TyCtxt<'tcx>,
    closure_instance: Instance<'tcx>,
)

For a given closure, writes out the data for the profiling the impact of RFC 2229 on closure size into a CSV.

During the same compile all closures dump the information in the same file “closure_profile_XXXXX.csv”, which is created in the directory where the compiler is invoked.