rustc_monomorphize::mono_checks
Module abi_check

This module ensures that if a function’s ABI requires a particular target feature, that target feature is enabled both on the callee and all callers.

Functions

check_call_site_abi 🔒

Checks that a call expression does not try to pass a vector-passed argument which requires a target feature that the caller does not have, as doing so causes UB because of ABI mismatch.

check_callees_abi 🔒

check_feature_dependent_abi 🔒

check_instance_abi 🔒

Checks that the ABI of a given instance of a function does not contain vector-passed arguments or return values for which the corresponding target feature is not enabled.

do_check_abi 🔒

Checks whether a certain function ABI is compatible with the target features currently enabled for a certain function. If not, emit_err is called, with Some(feature) if a certain feature should be enabled and with None if no feature is known that would make the ABI compatible.

uses_vector_registers 🔒